Refactor reducers task.
