+++ src/components/routeComponents/memberInformationGrid/markOptInOrOptOutSelectedMember/MarkOptInOrOptOutSelectedMember.jsx
@@ -22,9 +22,11 @@ import {
 } from 'actions/memberRegistrationActions';
 import {
   getSecretKey,
+} from 'reducers/memberRegistrationReducer';
+import {
   isMarkOptInOrOptOutSuccess,
   isMarkOptInOrOptOutFailed,
-} from 'reducers/memberRegistrationReducer';
+} from 'reducers/allMembersDataReducer';
 import {
   THIS_INFORMATION_IS_COMPULSORY_MESSAGE,
 } from 'constants/messages';
