+++ src/components/routeComponents/memberInformationGrid/updateIdCardStatusMembersModal/UpdateIdCardStatusMembersModal.jsx
@@ -21,9 +21,11 @@ import {
 } from 'actions/memberRegistrationActions';
 import {
   getSecretKey,
+} from 'reducers/memberRegistrationReducer';
+import {
   isUpdateIdCardStatusSuccess,
   isUpdateIdCardStatusFailed,
-} from 'reducers/memberRegistrationReducer';
+} from 'reducers/allMembersDataReducer';
 import {
   THIS_INFORMATION_IS_COMPULSORY_MESSAGE,
 } from 'constants/messages';
