+++ src/components/routeComponents/memberInformationGrid/uploadMembersAttendanceFile/UploadMembersAttendanceFile.jsx
@@ -22,11 +22,13 @@ import {
 } from 'actions/memberRegistrationActions';
 import {
   getSecretKey,
+} from 'reducers/memberRegistrationReducer';
+import {
   getSuccess,
   getFailRecordIds,
   isUploadAttendanceFailed,
   idNotExistErrorMessage,
-} from 'reducers/memberRegistrationReducer';
+} from 'reducers/allMembersDataReducer';
 import {
   THIS_INFORMATION_IS_COMPULSORY_MESSAGE,
 } from 'constants/messages';
