+++ src/sagas/index.js
@@ -6,7 +6,6 @@ import {
   markSelectedMembersAttendanceSaga,
   markSelectedMembersOptInOrOptOutSaga,
   parentsRegistrationSaga,
-  searchMemberSaga,
   updateIdCardStatusSelectedMembersSaga,
   updateMemberSaga,
   uploadAttendanceFileSaga,
@@ -18,7 +17,6 @@ const sagas = [
   takeLatest(['CREATE_MEMBER'], createMemberSaga),
   takeLatest(['FETCH_MEMBER'], fetchMemberSaga),
   takeLatest(['UPDATE_MEMBER'], updateMemberSaga),
-  takeLatest(['FETCH_SEARCH_RESULTS'], searchMemberSaga),
   takeLatest(['GET_ALL_MEMBERS'], getAllMembersSaga),
   takeLatest(['UPLOAD_ATTENDANCE_FILE'], uploadAttendanceFileSaga),
   takeLatest(['UPLOAD_OPT_IN_FILE'], uploadOptInFileSaga),
